requirements.txt
# 浏览器自动化 - 使用较新版本以支持 Python 3.13
playwright==1.48.0

# 配置文件处理
pyyaml==6.0.2

# 环境变量
python-dotenv==1.0.1

# 验证码服务
2captcha-python==1.2.2

# HTTP客户端 - 更新版本
aiohttp==3.10.11

# 日志处理
loguru==0.7.2

# 加密库 - 更新版本
cryptography==43.0.3

# 数据处理 - 更新版本
pandas==2.2.3

# 邮件处理
imapclient==3.0.1

# 时区处理
pytz==2024.1

# 代理支持
aiohttp-socks==0.9.0

# 异步支持
asyncio-throttle==1.0.2